Refactor previous name into dedicated service
Fixes #7568 by refactoring the patient previous name into its own class.
